refactor: 수익 계산 시 정밀도 보장을 위한 BigDecimal 적용

+++ src/main/java/blackjack/domain/gameresult/GameResults.java
@@ -19,7 +19,7 @@ public class GameResults {
         Map<Player, Integer> results = new LinkedHashMap<>();
         for (Player player : players.getPlayers()) {
             GameResult gameResult = GameResult.matchResult(player, dealer);
-            results.put(player, player.calculateProfit(gameResult.getRatio()));
+            results.put(player, player.calculateProfit(gameResult.getRatio()).intValue());
         }
         return new GameResults(results);
     }
